+++ src/lib/output/themes/default/partials/typeDetails.tsx
@@ -1,79 +1,142 @@
 import {
+    Comment,
     type CommentDisplayPart,
     type DeclarationReflection,
-    type Reflection,
+    Reflection,
     ReflectionKind,
     type SignatureReflection,
 } from "../../../../models/index.js";
 import type { ReferenceType, SomeType, TypeVisitor } from "../../../../models/types.js";
-import { i18n, JSX } from "#utils";
+import { assert, i18n, JSX } from "#utils";
 import { classNames, getKindClass } from "../../lib.js";
 import type { DefaultThemeRenderContext } from "../DefaultThemeRenderContext.js";
 import { anchorTargetIfPresent } from "./anchor-icon.js";
 
-const isUsefulVisitor: Partial<TypeVisitor<boolean>> = {
-    array(type) {
-        return renderingTypeDetailsIsUseful(type.elementType);
-    },
-    intersection(type) {
-        return type.types.some(renderingTypeDetailsIsUseful);
-    },
-    union(type) {
-        return !!type.elementSummaries || type.types.some(renderingTypeDetailsIsUseful);
-    },
-    reflection(type) {
-        return renderingChildIsUseful(type.declaration);
-    },
-    reference(type) {
-        return shouldExpandReference(type);
-    },
-};
+function renderingTypeDetailsIsUseful(container: Reflection, type: SomeType): boolean {
+    const isUsefulVisitor: Partial<TypeVisitor<boolean>> = {
+        array(type) {
+            return renderingTypeDetailsIsUseful(container, type.elementType);
+        },
+        intersection(type) {
+            return type.types.some(t => renderingTypeDetailsIsUseful(container, t));
+        },
+        union(type) {
+            return !!type.elementSummaries || type.types.some(t => renderingTypeDetailsIsUseful(container, t));
+        },
+        reflection(type) {
+            return renderingChildIsUseful(type.declaration);
+        },
+        reference(type) {
+            return shouldExpandReference(container, type);
+        },
+    };
 
-function renderingTypeDetailsIsUseful(type: SomeType) {
     return type.visit(isUsefulVisitor) ?? false;
 }
 
-export function typeDeclaration(context: DefaultThemeRenderContext, type: SomeType): JSX.Children {
-    if (renderingTypeDetailsIsUseful(type)) {
+export function typeDeclaration(
+    context: DefaultThemeRenderContext,
+    reflectionOwningType: Reflection,
+    type: SomeType,
+): JSX.Children {
+    assert(
+        reflectionOwningType instanceof Reflection,
+        "typeDeclaration(reflectionOwningType, type) called incorrectly",
+    );
+
+    if (renderingTypeDetailsIsUseful(reflectionOwningType, type)) {
         return (
             <div class="tsd-type-declaration">
                 <h4>{i18n.theme_type_declaration()}</h4>
-                {context.typeDetails(type, true)}
+                {context.typeDetails(reflectionOwningType, type, true)}
             </div>
         );
     }
     return null;
 }
 
-const expanded = new Set<Reflection>();
-function shouldExpandReference(reference: ReferenceType) {
-    const target = reference.reflection;
-    if (reference.highlightedProperties) {
-        return !target || expanded.has(target) === false;
+type ExpandTypeInfo = { expandType: Set<string>; preventExpand: Set<string> };
+const expandTypeCache = new WeakMap<Reflection, ExpandTypeInfo>();
+
+function getExpandTypeInfo(refl: Reflection): ExpandTypeInfo {
+    const cache = expandTypeCache.get(refl);
+    if (cache) return cache;
+
+    const expandType = new Set<string>();
+    const preventExpand = new Set<string>();
+    if (!refl.isProject()) {
+        const info = getExpandTypeInfo(refl.parent!);
+        for (const item of info.expandType) {
+            expandType.add(item);
+        }
+        for (const item of info.preventExpand) {
+            preventExpand.add(item);
+        }
     }
 
-    if (!target?.kindOf(ReflectionKind.TypeAlias | ReflectionKind.Interface)) return false;
-    if (!target.comment?.hasModifier("@expand")) return false;
+    for (const tag of refl.comment?.blockTags || []) {
+        if (tag.tag === "@expandType") {
+            const name = Comment.combineDisplayParts(tag.content);
+            expandType.add(name);
+            preventExpand.delete(name);
+        } else if (tag.tag === "@preventExpand") {
+            const name = Comment.combineDisplayParts(tag.content);
+            preventExpand.add(name);
+            expandType.delete(name);
+        }
+    }
 
-    return expanded.has(target) === false;
+    expandTypeCache.set(refl, { expandType, preventExpand });
+    return { expandType, preventExpand };
 }
 
-export function typeDetails(context: DefaultThemeRenderContext, type: SomeType, renderAnchors: boolean): JSX.Children {
-    return typeDetailsImpl(context, type, renderAnchors);
+const expanded = new Set<Reflection>();
+function shouldExpandReference(container: Reflection, reference: ReferenceType) {
+    const target = reference.reflection;
+    if (!target) {
+        // If it doesn't exist, expand only if there are specific properties
+        // which the user annotated. Assume they know what they're doing.
+        return reference.highlightedProperties !== undefined;
+    }
+
+    // Prevent expansion of non-types
+    if (!target.kindOf(ReflectionKind.TypeAlias | ReflectionKind.Interface)) return false;
+
+    // Prevent recursive expand
+    if (expanded.has(target)) return false;
+
+    const info = getExpandTypeInfo(container);
+
+    // Expand if the user explicitly requested it with @param or @expand
+    if (reference.highlightedProperties || target.comment?.hasModifier("@expand") || info.expandType.has(target.name)) {
+        return !info.preventExpand.has(target.name);
+    }
+
+    return false;
+}
+
+export function typeDetails(
+    context: DefaultThemeRenderContext,
+    reflectionOwningType: Reflection,
+    type: SomeType,
+    renderAnchors: boolean,
+): JSX.Children {
+    return typeDetailsImpl(context, reflectionOwningType, type, renderAnchors);
 }
 
 export function typeDetailsImpl(
     context: DefaultThemeRenderContext,
+    reflectionOwningType: Reflection,
     type: SomeType,
     renderAnchors: boolean,
     highlighted?: Map<string, CommentDisplayPart[]>,
 ): JSX.Children {
     const result = type.visit<JSX.Children>({
         array(type) {
-            return context.typeDetails(type.elementType, renderAnchors);
+            return context.typeDetails(reflectionOwningType, type.elementType, renderAnchors);
         },
         intersection(type) {
-            return type.types.map((t) => context.typeDetails(t, renderAnchors));
+            return type.types.map((t) => context.typeDetails(reflectionOwningType, t, renderAnchors));
         },
         union(type) {
             const result: JSX.Children = [];
@@ -82,7 +145,7 @@ export function typeDetailsImpl(
                     <li>
                         {context.type(type.types[i])}
                         {context.displayParts(type.elementSummaries?.[i])}
-                        {context.typeDetailsIfUseful(type.types[i])}
+                        {context.typeDetailsIfUseful(reflectionOwningType, type.types[i])}
                     </li>,
                 );
             }
@@ -96,7 +159,7 @@ export function typeDetailsImpl(
             return declarationDetails(context, declaration, renderAnchors);
         },
         reference(reference) {
-            if (shouldExpandReference(reference)) {
+            if (shouldExpandReference(reflectionOwningType, reference)) {
                 const target = reference.reflection;
                 if (!target?.isDeclaration()) {
                     return highlightedPropertyDetails(context, reference.highlightedProperties);
@@ -105,7 +168,7 @@ export function typeDetailsImpl(
                 // Ensure we don't go into an infinite loop here
                 expanded.add(target);
                 const details = target.type
-                    ? context.typeDetails(target.type, renderAnchors)
+                    ? context.typeDetails(reflectionOwningType, target.type, renderAnchors)
                     : declarationDetails(context, target, renderAnchors);
                 expanded.delete(target);
                 return details;
@@ -121,9 +184,18 @@ export function typeDetailsImpl(
     return result;
 }
 
-export function typeDetailsIfUseful(context: DefaultThemeRenderContext, type: SomeType | undefined): JSX.Children {
-    if (type && renderingTypeDetailsIsUseful(type)) {
-        return context.typeDetails(type, false);
+export function typeDetailsIfUseful(
+    context: DefaultThemeRenderContext,
+    reflectionOwningType: Reflection,
+    type: SomeType | undefined,
+): JSX.Children {
+    assert(
+        reflectionOwningType instanceof Reflection,
+        "typeDetailsIfUseful(reflectionOwningType, type) called incorrectly",
+    );
+
+    if (type && renderingTypeDetailsIsUseful(reflectionOwningType, type)) {
+        return context.typeDetails(reflectionOwningType, type, false);
     }
 }
 
@@ -324,7 +396,7 @@ function renderIndexSignature(context: DefaultThemeRenderContext, index: Signatu
             </h5>
             {context.commentSummary(index)}
             {context.commentTags(index)}
-            {context.typeDeclaration(index.type!)}
+            {context.typeDeclaration(index, index.type!)}
         </li>
     );
 }